Change package name from `projections` to `projection`.
