+++ api-java/src/main/java/org/spine3/examples/todolist/q/projection/DraftTasksViewProjection.java
@@ -18,25 +18,19 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.examples.todolist.q.projections;
+package org.spine3.examples.todolist.q.projection;
 
-import org.spine3.base.EventContext;
-import org.spine3.examples.todolist.TaskDefinition;
 import org.spine3.examples.todolist.TaskDetails;
-import org.spine3.examples.todolist.TaskId;
 import org.spine3.examples.todolist.TaskListId;
-import org.spine3.examples.todolist.c.enrichments.TaskEnrichment;
 import org.spine3.examples.todolist.c.events.LabelAssignedToTask;
 import org.spine3.examples.todolist.c.events.LabelDetailsUpdated;
 import org.spine3.examples.todolist.c.events.LabelRemovedFromTask;
-import org.spine3.examples.todolist.c.events.TaskCompleted;
-import org.spine3.examples.todolist.c.events.TaskCreated;
 import org.spine3.examples.todolist.c.events.TaskDeleted;
 import org.spine3.examples.todolist.c.events.TaskDescriptionUpdated;
+import org.spine3.examples.todolist.c.events.TaskDraftCreated;
 import org.spine3.examples.todolist.c.events.TaskDraftFinalized;
 import org.spine3.examples.todolist.c.events.TaskDueDateUpdated;
 import org.spine3.examples.todolist.c.events.TaskPriorityUpdated;
-import org.spine3.examples.todolist.c.events.TaskReopened;
 import org.spine3.server.event.Subscribe;
 import org.spine3.server.projection.Projection;
 
@@ -44,23 +38,20 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import static org.spine3.base.Identifiers.newUuid;
-import static org.spine3.examples.todolist.EnrichmentHelper.getEnrichment;
-import static org.spine3.examples.todolist.q.projections.ProjectionHelper.removeViewByTaskId;
-import static org.spine3.examples.todolist.q.projections.ProjectionHelper.updateTaskViewList;
+import static org.spine3.examples.todolist.q.projection.ProjectionHelper.removeViewByTaskId;
+import static org.spine3.examples.todolist.q.projection.ProjectionHelper.updateTaskViewList;
 
 /**
- * A projections state of the finalized tasks.
+ * A projection state of tasks in a "draft" state.
  *
- * <p> Contains the task list view items.
- * <p> This view includes all tasks that are not in a draft state and not deleted.
+ * <p> This view includes all non-deleted tasks, which creation is in progress.
  *
  * @author Illia Shepilov
  */
-@SuppressWarnings("OverlyCoupledClass")
-public class MyListViewProjection extends Projection<TaskListId, MyListView> {
+public class DraftTasksViewProjection extends Projection<TaskListId, DraftTasksView> {
 
     /**
-     * As long as there is just a single end-user in the app, the {@link MyListViewProjection} is a singleton.
+     * As long as there is just a single end-user in the app, the {@link DraftTasksViewProjection} is a singleton.
      */
     public static final TaskListId ID = TaskListId.newBuilder()
                                                   .setValue(newUuid())
@@ -72,12 +63,12 @@ public class MyListViewProjection extends Projection<TaskListId, MyListView> {
      * @param id the ID for the new instance
      * @throws IllegalArgumentException if the ID is not of one of the supported types
      */
-    public MyListViewProjection(TaskListId id) {
+    public DraftTasksViewProjection(TaskListId id) {
         super(id);
     }
 
     @Subscribe
-    public void on(TaskCreated event) {
+    public void on(TaskDraftCreated event) {
         final TaskDetails taskDetails = event.getDetails();
         final TaskView taskView = TaskView.newBuilder()
                                           .setId(event.getId())
@@ -85,127 +76,107 @@ public class MyListViewProjection extends Projection<TaskListId, MyListView> {
                                           .setPriority(taskDetails.getPriority())
                                           .setCompleted(taskDetails.getCompleted())
                                           .build();
-        final List<TaskView> views = getState().getMyList()
+        final List<TaskView> views = getState().getDraftTasks()
                                                .getItemsList()
                                                .stream()
                                                .collect(Collectors.toList());
         views.add(taskView);
-        final MyListView state = constructMyListViewState(views);
+        final TaskListView taskListView = TaskListView.newBuilder()
+                                                      .addAllItems(views)
+                                                      .build();
+        final DraftTasksView state = getState().newBuilderForType()
+                                               .setDraftTasks(taskListView)
+                                               .build();
+        incrementState(state);
+    }
+
+    @Subscribe
+    public void on(TaskDraftFinalized event) {
+        final List<TaskView> views = getState().getDraftTasks()
+                                               .getItemsList()
+                                               .stream()
+                                               .collect(Collectors.toList());
+        final TaskListView taskListView = removeViewByTaskId(views, event.getTaskId());
+        final DraftTasksView state = getState().newBuilderForType()
+                                               .setDraftTasks(taskListView)
+                                               .build();
         incrementState(state);
     }
 
     @Subscribe
     public void on(TaskDeleted event) {
-        final List<TaskView> views = getState().getMyList()
+        final List<TaskView> views = getState().getDraftTasks()
                                                .getItemsList()
                                                .stream()
                                                .collect(Collectors.toList());
         final TaskListView taskListView = removeViewByTaskId(views, event.getTaskId());
-        final MyListView state = getState().newBuilderForType()
-                                           .setMyList(taskListView)
-                                           .build();
+        final DraftTasksView state = getState().newBuilderForType()
+                                               .setDraftTasks(taskListView)
+                                               .build();
         incrementState(state);
     }
 
     @Subscribe
     public void on(TaskDescriptionUpdated event) {
-        final List<TaskView> views = getState().getMyList()
+        final List<TaskView> views = getState().getDraftTasks()
                                                .getItemsList();
         final List<TaskView> updatedList = updateTaskViewList(views, event);
-        final MyListView state = constructMyListViewState(updatedList);
+        final DraftTasksView state = constructDraftTasksViewState(updatedList);
         incrementState(state);
     }
 
     @Subscribe
     public void on(TaskPriorityUpdated event) {
-        final List<TaskView> views = getState().getMyList()
+        final List<TaskView> views = getState().getDraftTasks()
                                                .getItemsList();
         final List<TaskView> updatedList = updateTaskViewList(views, event);
-        final MyListView state = constructMyListViewState(updatedList);
+        final DraftTasksView state = constructDraftTasksViewState(updatedList);
         incrementState(state);
     }
 
     @Subscribe
     public void on(TaskDueDateUpdated event) {
-        final List<TaskView> views = getState().getMyList()
+        final List<TaskView> views = getState().getDraftTasks()
                                                .getItemsList();
         final List<TaskView> updatedList = updateTaskViewList(views, event);
-        final MyListView state = constructMyListViewState(updatedList);
-        incrementState(state);
-    }
-
-    @Subscribe
-    public void on(TaskCompleted event) {
-        final List<TaskView> views = getState().getMyList()
-                                               .getItemsList();
-        final List<TaskView> updatedList = updateTaskViewList(views, event);
-        final MyListView state = constructMyListViewState(updatedList);
-        incrementState(state);
-    }
-
-    @Subscribe
-    public void on(TaskReopened event) {
-        final List<TaskView> views = getState().getMyList()
-                                               .getItemsList();
-        final List<TaskView> updatedList = updateTaskViewList(views, event);
-        final MyListView state = constructMyListViewState(updatedList);
+        final DraftTasksView state = constructDraftTasksViewState(updatedList);
         incrementState(state);
     }
 
     @Subscribe
     public void on(LabelAssignedToTask event) {
-        final List<TaskView> views = getState().getMyList()
+        final List<TaskView> views = getState().getDraftTasks()
                                                .getItemsList();
         final List<TaskView> updatedList = updateTaskViewList(views, event);
-        final MyListView state = constructMyListViewState(updatedList);
+        final DraftTasksView state = constructDraftTasksViewState(updatedList);
         incrementState(state);
     }
 
     @Subscribe
     public void on(LabelRemovedFromTask event) {
-        final List<TaskView> views = getState().getMyList()
+        final List<TaskView> views = getState().getDraftTasks()
                                                .getItemsList();
         final List<TaskView> updatedList = updateTaskViewList(views, event);
-        final MyListView state = constructMyListViewState(updatedList);
+        final DraftTasksView state = constructDraftTasksViewState(updatedList);
         incrementState(state);
     }
 
     @Subscribe
     public void on(LabelDetailsUpdated event) {
-        final List<TaskView> views = getState().getMyList()
+        final List<TaskView> views = getState().getDraftTasks()
                                                .getItemsList();
         final List<TaskView> updatedList = updateTaskViewList(views, event);
-        final MyListView state = constructMyListViewState(updatedList);
+        final DraftTasksView state = constructDraftTasksViewState(updatedList);
         incrementState(state);
     }
 
-    @Subscribe
-    public void on(TaskDraftFinalized event, EventContext context) {
-        final TaskId taskId = event.getTaskId();
-        final TaskEnrichment enrichment = getEnrichment(TaskEnrichment.class, context);
-        final TaskDefinition taskDefinition = enrichment.getTaskDefinition();
-        final TaskView view = TaskView.newBuilder()
-                                      .setId(taskId)
-                                      .setDescription(taskDefinition.getDescription())
-                                      .setDueDate(taskDefinition.getDueDate())
-                                      .setPriority(taskDefinition.getPriority())
-                                      .build();
-        final List<TaskView> views = getState().getMyList()
-                                               .getItemsList()
-                                               .stream()
-                                               .collect(Collectors.toList());
-        views.add(view);
-        final MyListView state = constructMyListViewState(views);
-        incrementState(state);
-    }
-
-    private MyListView constructMyListViewState(List<TaskView> updatedList) {
+    private DraftTasksView constructDraftTasksViewState(List<TaskView> updatedList) {
         final TaskListView listView = TaskListView.newBuilder()
                                                   .addAllItems(updatedList)
                                                   .build();
-        final MyListView result = getState().newBuilderForType()
-                                            .setMyList(listView)
-                                            .build();
+        final DraftTasksView result = getState().newBuilderForType()
+                                                .setDraftTasks(listView)
+                                                .build();
         return result;
     }
 }
