+++ api-java/src/test/java/org/spine3/examples/todolist/q/projection/DraftTasksViewProjectionTest.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.examples.todolist.q.projections;
+package org.spine3.examples.todolist.q.projection;
 
 import com.google.protobuf.Timestamp;
 import org.junit.jupiter.api.BeforeEach;
@@ -32,19 +32,18 @@ import org.spine3.examples.todolist.TaskListId;
 import org.spine3.examples.todolist.TaskPriority;
 import org.spine3.examples.todolist.c.events.LabelAssignedToTask;
 import org.spine3.examples.todolist.c.events.LabelDetailsUpdated;
-import org.spine3.examples.todolist.c.events.TaskCompleted;
-import org.spine3.examples.todolist.c.events.TaskCreated;
+import org.spine3.examples.todolist.c.events.LabelRemovedFromTask;
 import org.spine3.examples.todolist.c.events.TaskDeleted;
 import org.spine3.examples.todolist.c.events.TaskDescriptionUpdated;
+import org.spine3.examples.todolist.c.events.TaskDraftCreated;
+import org.spine3.examples.todolist.c.events.TaskDraftFinalized;
 import org.spine3.examples.todolist.c.events.TaskDueDateUpdated;
 import org.spine3.examples.todolist.c.events.TaskPriorityUpdated;
-import org.spine3.examples.todolist.c.events.TaskReopened;
 import org.spine3.protobuf.Timestamps;
 
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.spine3.examples.todolist.testdata.TestEventFactory.DESCRIPTION;
@@ -54,49 +53,72 @@ import static org.spine3.examples.todolist.testdata.TestEventFactory.TASK_PRIORI
 import static org.spine3.examples.todolist.testdata.TestEventFactory.UPDATED_DESCRIPTION;
 import static org.spine3.examples.todolist.testdata.TestEventFactory.labelAssignedToTaskInstance;
 import static org.spine3.examples.todolist.testdata.TestEventFactory.labelDetailsUpdatedInstance;
-import static org.spine3.examples.todolist.testdata.TestEventFactory.taskCompletedInstance;
-import static org.spine3.examples.todolist.testdata.TestEventFactory.taskCreatedInstance;
+import static org.spine3.examples.todolist.testdata.TestEventFactory.labelRemovedFromTaskInstance;
 import static org.spine3.examples.todolist.testdata.TestEventFactory.taskDeletedInstance;
 import static org.spine3.examples.todolist.testdata.TestEventFactory.taskDescriptionUpdatedInstance;
+import static org.spine3.examples.todolist.testdata.TestEventFactory.taskDraftCreatedInstance;
+import static org.spine3.examples.todolist.testdata.TestEventFactory.taskDraftFinalizedInstance;
 import static org.spine3.examples.todolist.testdata.TestEventFactory.taskDueDateUpdatedInstance;
 import static org.spine3.examples.todolist.testdata.TestEventFactory.taskPriorityUpdatedInstance;
-import static org.spine3.examples.todolist.testdata.TestEventFactory.taskReopenedInstance;
 import static org.spine3.examples.todolist.testdata.TestTaskLabelCommandFactory.UPDATED_LABEL_TITLE;
 
 /**
  * @author Illia Shepilov
  */
-public class MyListViewProjectionTest extends ProjectionTest {
+@SuppressWarnings("OverlyCoupledClass")
+public class DraftTasksViewProjectionTest extends ProjectionTest {
 
-    private MyListViewProjection projection;
+    private DraftTasksViewProjection projection;
 
     @BeforeEach
-    void setUp() {
+    public void setUp() {
         final TaskListId taskListId = createTaskListId();
-        projection = new MyListViewProjection(taskListId);
+        projection = new DraftTasksViewProjection(taskListId);
+    }
+
+    private void taskDraftCreated() {
+        final TaskDraftCreated taskDraftCreatedEvent = taskDraftCreatedInstance();
+        projection.on(taskDraftCreatedEvent);
     }
 
     @Nested
-    @DisplayName("TaskCreated event")
-    class TaskCreatedEvent {
+    @DisplayName("TaskDraftCreated event")
+    class TaskDraftCreatedEvent {
 
         @Test
         @DisplayName(ADDS_TASK_VIEW_TO_STATE)
         public void addsView() {
-            final TaskCreated taskCreatedEvent = taskCreatedInstance();
-            final int expectedSize = 1;
-            projection.on(taskCreatedEvent);
+            taskDraftCreated();
+
+            final TaskListView listView = projection.getState()
+                                                    .getDraftTasks();
+            final int expectedListSize = 1;
+            assertEquals(expectedListSize, listView.getItemsCount());
+
+            final TaskView taskView = listView.getItems(0);
+
+            assertEquals(TASK_ID, taskView.getId());
+            assertEquals(DESCRIPTION, taskView.getDescription());
+            assertEquals(TASK_PRIORITY, taskView.getPriority());
+        }
+    }
+
+    @Nested
+    @DisplayName("TaskDraftFinalized event")
+    class TaskDraftFinalizedEvent {
+
+        @Test
+        @DisplayName(REMOVES_TASK_VIEW_FORM_STATE)
+        public void removesView() {
+            taskDraftCreated();
+
+            final TaskDraftFinalized taskDraftFinalizedEvent = taskDraftFinalizedInstance();
+            projection.on(taskDraftFinalizedEvent);
 
             final List<TaskView> views = projection.getState()
-                                                   .getMyList()
+                                                   .getDraftTasks()
                                                    .getItemsList();
-
-            assertEquals(expectedSize, views.size());
-
-            final TaskView view = views.get(0);
-
-            assertEquals(TASK_PRIORITY, view.getPriority());
-            assertEquals(DESCRIPTION, view.getDescription());
+            assertTrue(views.isEmpty());
         }
     }
 
@@ -107,23 +129,14 @@ public class MyListViewProjectionTest extends ProjectionTest {
         @Test
         @DisplayName(REMOVES_TASK_VIEW_FORM_STATE)
         public void removesView() {
-            final TaskCreated taskCreatedEvent = taskCreatedInstance();
-            projection.on(taskCreatedEvent);
-            projection.on(taskCreatedEvent);
+            taskDraftCreated();
 
             final TaskDeleted taskDeletedEvent = taskDeletedInstance();
             projection.on(taskDeletedEvent);
 
-            List<TaskView> views = projection.getState()
-                                             .getMyList()
-                                             .getItemsList();
-            final int expectedListSize = 1;
-            assertEquals(expectedListSize, views.size());
-
-            projection.on(taskDeletedEvent);
-            views = projection.getState()
-                              .getMyList()
-                              .getItemsList();
+            final List<TaskView> views = projection.getState()
+                                                   .getDraftTasks()
+                                                   .getItemsList();
             assertTrue(views.isEmpty());
         }
     }
@@ -135,17 +148,16 @@ public class MyListViewProjectionTest extends ProjectionTest {
         @Test
         @DisplayName(UPDATES_TASK_DESCRIPTION)
         public void updatesDescription() {
-            final TaskCreated taskCreatedEvent = taskCreatedInstance();
-            projection.on(taskCreatedEvent);
+            final TaskDraftCreated taskDraftCreatedEvent = taskDraftCreatedInstance();
+            projection.on(taskDraftCreatedEvent);
 
-            final TaskId expectedTaskId = taskCreatedEvent.getId();
+            final TaskId expectedTaskId = taskDraftCreatedEvent.getId();
             final TaskDescriptionUpdated descriptionUpdatedEvent =
                     taskDescriptionUpdatedInstance(expectedTaskId, UPDATED_DESCRIPTION);
             projection.on(descriptionUpdatedEvent);
 
             final TaskListView taskListView = projection.getState()
-                                                        .getMyList();
-
+                                                        .getDraftTasks();
             final int expectedViewSize = 1;
             assertEquals(expectedViewSize, taskListView.getItemsCount());
 
@@ -158,8 +170,7 @@ public class MyListViewProjectionTest extends ProjectionTest {
         @Test
         @DisplayName(DOES_NOT_UPDATE_TASK_DESCRIPTION_BY_WRONG_TASK_ID)
         public void doesNotUpdateDescription() {
-            final TaskCreated taskCreatedEvent = taskCreatedInstance();
-            projection.on(taskCreatedEvent);
+            taskDraftCreated();
 
             final String updatedDescription = UPDATED_DESCRIPTION;
 
@@ -168,7 +179,7 @@ public class MyListViewProjectionTest extends ProjectionTest {
             projection.on(descriptionUpdatedEvent);
 
             final TaskListView taskListView = projection.getState()
-                                                        .getMyList();
+                                                        .getDraftTasks();
             final int expectedViewListSize = 1;
             assertEquals(expectedViewListSize, taskListView.getItemsCount());
 
@@ -181,22 +192,22 @@ public class MyListViewProjectionTest extends ProjectionTest {
 
     @Nested
     @DisplayName(TASK_DUE_DATE_UPDATED_EVENT)
-    class TaskDueDateUpdatedEvent {
+    class DueDateUpdatedEvent {
 
         @Test
         @DisplayName(UPDATES_TASK_DUE_DATE)
         public void updatesDueDate() {
-            final TaskCreated taskCreatedEvent = taskCreatedInstance();
-            projection.on(taskCreatedEvent);
+            final TaskDraftCreated taskDraftCreatedEvent = taskDraftCreatedInstance();
+            projection.on(taskDraftCreatedEvent);
 
             final Timestamp updatedDueDate = Timestamps.getCurrentTime();
-            final TaskId expectedTaskId = taskCreatedEvent.getId();
+            final TaskId expectedTaskId = taskDraftCreatedEvent.getId();
 
             final TaskDueDateUpdated taskDueDateUpdatedEvent = taskDueDateUpdatedInstance(expectedTaskId, updatedDueDate);
             projection.on(taskDueDateUpdatedEvent);
 
             final TaskListView taskListView = projection.getState()
-                                                        .getMyList();
+                                                        .getDraftTasks();
             final int expectedViewListSize = 1;
             assertEquals(expectedViewListSize, taskListView.getItemsCount());
 
@@ -207,10 +218,9 @@ public class MyListViewProjectionTest extends ProjectionTest {
         }
 
         @Test
-        @DisplayName(DOES_NOT_UPDATE_TASK_DUE_DATE)
-        public void doeNotUpdateDueDate() {
-            final TaskCreated taskCreatedEvent = taskCreatedInstance();
-            projection.on(taskCreatedEvent);
+        @DisplayName("does not update task due date by wrong task ID")
+        public void doesNotUpdate() {
+            taskDraftCreated();
 
             final Timestamp updatedDueDate = Timestamps.getCurrentTime();
 
@@ -219,12 +229,13 @@ public class MyListViewProjectionTest extends ProjectionTest {
             projection.on(taskDueDateUpdatedEvent);
 
             final TaskListView taskListView = projection.getState()
-                                                        .getMyList();
+                                                        .getDraftTasks();
             final int expectedViewListSize = 1;
             assertEquals(expectedViewListSize, taskListView.getItemsCount());
 
             final TaskView taskView = taskListView.getItemsList()
                                                   .get(0);
+            assertEquals(TASK_ID, taskView.getId());
             assertNotEquals(updatedDueDate, taskView.getDueDate());
         }
     }
@@ -236,18 +247,18 @@ public class MyListViewProjectionTest extends ProjectionTest {
         @Test
         @DisplayName(UPDATES_TASK_PRIORITY)
         public void updatesPriority() {
-            final TaskCreated taskCreatedEvent = taskCreatedInstance();
-            projection.on(taskCreatedEvent);
+            final TaskDraftCreated taskDraftCreatedEvent = taskDraftCreatedInstance();
+            projection.on(taskDraftCreatedEvent);
 
             final TaskPriority updatedTaskPriority = TaskPriority.LOW;
-            final TaskId expectedTaskId = taskCreatedEvent.getId();
+            final TaskId expectedTaskId = taskDraftCreatedEvent.getId();
 
             final TaskPriorityUpdated taskPriorityUpdatedEvent =
                     taskPriorityUpdatedInstance(expectedTaskId, updatedTaskPriority);
             projection.on(taskPriorityUpdatedEvent);
 
             final TaskListView taskListView = projection.getState()
-                                                        .getMyList();
+                                                        .getDraftTasks();
             final int expectedViewListSize = 1;
             assertEquals(expectedViewListSize, taskListView.getItemsCount());
 
@@ -260,8 +271,7 @@ public class MyListViewProjectionTest extends ProjectionTest {
         @Test
         @DisplayName(DOES_NOT_UPDATE_TASK_PRIORITY_BY_WRONG_TASK_ID)
         public void doesNotUpdatePriority() {
-            final TaskCreated taskCreatedEvent = taskCreatedInstance();
-            projection.on(taskCreatedEvent);
+            taskDraftCreated();
 
             final TaskPriority updatedTaskPriority = TaskPriority.LOW;
 
@@ -270,7 +280,7 @@ public class MyListViewProjectionTest extends ProjectionTest {
             projection.on(taskPriorityUpdatedEvent);
 
             final TaskListView taskListView = projection.getState()
-                                                        .getMyList();
+                                                        .getDraftTasks();
             final int expectedViewListSize = 1;
             assertEquals(expectedViewListSize, taskListView.getItemsCount());
 
@@ -288,10 +298,10 @@ public class MyListViewProjectionTest extends ProjectionTest {
         @Test
         @DisplayName(UPDATES_LABEL_DETAILS)
         public void updatesLabelDetails() {
-            final TaskCreated taskCreatedEvent = taskCreatedInstance();
-            projection.on(taskCreatedEvent);
+            final TaskDraftCreated taskDraftCreatedEvent = taskDraftCreatedInstance();
+            projection.on(taskDraftCreatedEvent);
 
-            final TaskId expectedTaskId = taskCreatedEvent.getId();
+            final TaskId expectedTaskId = taskDraftCreatedEvent.getId();
 
             final LabelAssignedToTask labelAssignedToTaskEvent = labelAssignedToTaskInstance(expectedTaskId, LABEL_ID);
             projection.on(labelAssignedToTaskEvent);
@@ -303,7 +313,7 @@ public class MyListViewProjectionTest extends ProjectionTest {
             projection.on(labelDetailsUpdatedEvent);
 
             final TaskListView taskListView = projection.getState()
-                                                        .getMyList();
+                                                        .getDraftTasks();
             final int expectedViewListSize = 1;
             assertEquals(expectedViewListSize, taskListView.getItemsCount());
 
@@ -316,10 +326,10 @@ public class MyListViewProjectionTest extends ProjectionTest {
         @Test
         @DisplayName(DOES_NOT_UPDATE_LABEL_DETAILS)
         public void doesNotUpdateLabelDetails() {
-            final TaskCreated taskCreatedEvent = taskCreatedInstance();
-            projection.on(taskCreatedEvent);
+            final TaskDraftCreated taskDraftCreatedEvent = taskDraftCreatedInstance();
+            projection.on(taskDraftCreatedEvent);
 
-            final TaskId expectedTaskId = taskCreatedEvent.getId();
+            final TaskId expectedTaskId = taskDraftCreatedEvent.getId();
 
             final LabelAssignedToTask labelAssignedToTaskEvent =
                     labelAssignedToTaskInstance(expectedTaskId, LabelId.getDefaultInstance());
@@ -332,110 +342,41 @@ public class MyListViewProjectionTest extends ProjectionTest {
             projection.on(labelDetailsUpdatedEvent);
 
             final TaskListView taskListView = projection.getState()
-                                                        .getMyList();
+                                                        .getDraftTasks();
             final int expectedViewListSize = 1;
             assertEquals(expectedViewListSize, taskListView.getItemsCount());
 
             final TaskView taskView = taskListView.getItemsList()
                                                   .get(0);
-            assertEquals(expectedTaskId, taskView.getId());
+            assertEquals(TASK_ID, taskView.getId());
             assertNotEquals(updatedColor, taskView.getLabelColor());
         }
     }
 
     @Nested
-    @DisplayName(TASK_COMPLETED_EVENT)
-    class TaskCompletedEvent {
+    @DisplayName(LABEL_REMOVED_FROM_TASK_EVENT)
+    class LabelRemovedFromTaskEvent {
 
         @Test
-        @DisplayName("completes task")
-        public void completesTask() {
-            final TaskCreated taskCreatedEvent = taskCreatedInstance();
-            projection.on(taskCreatedEvent);
+        @DisplayName(REMOVES_TASK_VIEW_FORM_STATE)
+        public void removesView() {
+            taskDraftCreated();
 
-            final TaskCompleted taskCompletedEvent = taskCompletedInstance();
-            projection.on(taskCompletedEvent);
+            final LabelAssignedToTask labelAssignedToTaskEvent = labelAssignedToTaskInstance();
+            projection.on(labelAssignedToTaskEvent);
+
+            final LabelRemovedFromTask labelRemovedFromTaskEvent = labelRemovedFromTaskInstance();
+            projection.on(labelRemovedFromTaskEvent);
 
             final TaskListView taskListView = projection.getState()
-                                                        .getMyList();
+                                                        .getDraftTasks();
             final int expectedViewListSize = 1;
             assertEquals(expectedViewListSize, taskListView.getItemsCount());
 
             final TaskView taskView = taskListView.getItemsList()
                                                   .get(0);
             assertEquals(TASK_ID, taskView.getId());
-            assertTrue(taskView.getCompleted());
-        }
-
-        @Test
-        @DisplayName("does not complete task by wrong task ID")
-        public void doesNotComplete() {
-            final TaskCreated taskCreatedEvent = taskCreatedInstance();
-            projection.on(taskCreatedEvent);
-
-            final TaskCompleted taskCompletedEvent = taskCompletedInstance(TaskId.getDefaultInstance());
-            projection.on(taskCompletedEvent);
-
-            final TaskListView taskListView = projection.getState()
-                                                        .getMyList();
-            final int expectedViewListSize = 1;
-            assertEquals(expectedViewListSize, taskListView.getItemsCount());
-
-            final TaskView taskView = taskListView.getItemsList()
-                                                  .get(0);
-            assertEquals(TASK_ID, taskView.getId());
-            assertFalse(taskView.getCompleted());
-        }
-    }
-
-    @Nested
-    @DisplayName(TASK_REOPENED_EVENT)
-    class TaskReopenedEvent {
-
-        @Test
-        @DisplayName("reopens task")
-        public void reopensTask() {
-            final TaskCreated taskCreatedEvent = taskCreatedInstance();
-            projection.on(taskCreatedEvent);
-
-            final TaskCompleted taskCompletedEvent = taskCompletedInstance();
-            projection.on(taskCompletedEvent);
-
-            final TaskReopened taskReopenedEvent = taskReopenedInstance();
-            projection.on(taskReopenedEvent);
-
-            final TaskListView taskListView = projection.getState()
-                                                        .getMyList();
-            final int expectedViewListSize = 1;
-            assertEquals(expectedViewListSize, taskListView.getItemsCount());
-
-            final TaskView taskView = taskListView.getItemsList()
-                                                  .get(0);
-            assertEquals(TASK_ID, taskView.getId());
-            assertFalse(taskView.getCompleted());
-        }
-
-        @Test
-        @DisplayName("does not reopen task by wrong task ID")
-        public void doesNotReopenTask() {
-            final TaskCreated taskCreatedEvent = taskCreatedInstance();
-            projection.on(taskCreatedEvent);
-
-            final TaskCompleted taskCompletedEvent = taskCompletedInstance();
-            projection.on(taskCompletedEvent);
-
-            final TaskReopened taskReopenedEvent = taskReopenedInstance(TaskId.getDefaultInstance());
-            projection.on(taskReopenedEvent);
-
-            final TaskListView taskListView = projection.getState()
-                                                        .getMyList();
-            final int expectedViewListSize = 1;
-            assertEquals(expectedViewListSize, taskListView.getItemsCount());
-
-            final TaskView taskView = taskListView.getItemsList()
-                                                  .get(0);
-            assertEquals(TASK_ID, taskView.getId());
-            assertTrue(taskView.getCompleted());
+            assertNotEquals(LABEL_ID, taskView.getLabelId());
         }
     }
 }
