+++ client/src/test/java/org/spine3/examples/todolist/client/CreateBasicTaskTest.java
@@ -24,8 +24,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.spine3.examples.todolist.TaskId;
 import org.spine3.examples.todolist.c.commands.CreateBasicTask;
-import org.spine3.examples.todolist.q.projections.LabelledTasksView;
-import org.spine3.examples.todolist.q.projections.TaskView;
+import org.spine3.examples.todolist.q.projection.LabelledTasksView;
+import org.spine3.examples.todolist.q.projection.TaskView;
 
 import java.util.ArrayList;
 import java.util.List;
