+++ client/src/test/java/org/spine3/examples/todolist/client/FinalizeDraftTest.java
@@ -29,9 +29,9 @@ import org.spine3.examples.todolist.c.commands.AssignLabelToTask;
 import org.spine3.examples.todolist.c.commands.CreateBasicLabel;
 import org.spine3.examples.todolist.c.commands.CreateDraft;
 import org.spine3.examples.todolist.c.commands.FinalizeDraft;
-import org.spine3.examples.todolist.q.projections.DraftTasksView;
-import org.spine3.examples.todolist.q.projections.LabelledTasksView;
-import org.spine3.examples.todolist.q.projections.TaskView;
+import org.spine3.examples.todolist.q.projection.DraftTasksView;
+import org.spine3.examples.todolist.q.projection.LabelledTasksView;
+import org.spine3.examples.todolist.q.projection.TaskView;
 
 import java.util.List;
